add in helper header for deducing atomic-ness of types
simple OR between two STL integral constants, but it is helpful to
reduce code duplication and will make it easier to patch if more atomic
types from HighFive want to be supported

+++ include/fire/EventHeader.hpp
@@ -135,7 +135,7 @@ class EventHeader {
   /**
    * The event timestamp
    */
-  int timestamp_;
+  long int timestamp_;
 
   /**
    * The event weight.
